refactor: Externalize DragAndDrop class, #14
